# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Change": {
				"enum": [
					"DESC",
					"ASC"
				],
				"type": "string"
			},
...
			"IProductReport": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"sales": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"brand",
					"quantity",
					"sales"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductReports": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/IProductReport"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"products"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProductSearchProps": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"sku": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"overallReview": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"imgPaths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"brand": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"createdAt",
					"sku",
					"name",
					"overallReview",
					"price",
					"imgPaths",
					"brand",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProducts": {
				"properties": {
					"products": {
						"items": {
							"$ref": "#/components/schemas/IProductSearchProps"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"products",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductField": {
				"enum": [
					"price",
					"name",
					"overallReview",
					"createdAt"
				],
				"type": "string"
			},
			"IVariantInfo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"properties": {
							"code": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"code",
							"name",
							"id"
						],
						"type": "object"
					},
					"size": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"productId",
					"quantity",
					"color",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductInfo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"sku": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"overallReview": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"imgPaths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"brand": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"summary": {
						"type": "string"
					},
					"variants": {
						"items": {
							"$ref": "#/components/schemas/IVariantInfo"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"sku",
					"name",
					"overallReview",
					"price",
					"imgPaths",
					"brand",
					"category",
					"summary",
					"variants"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateProductModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"sku": {
						"type": "string"
					},
					"categoryId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"brandId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"imgPaths": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"minItems": 0
					},
					"summary": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double",
						"minimum": 0
					}
				},
				"required": [
					"name",
					"categoryId",
					"brandId",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateVariantCreateModel": {
				"properties": {
					"sizeId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"colorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"quantity": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"productId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"sizeId",
					"colorId",
					"quantity",
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateVariantUpdateModel": {
				"properties": {
					"sizeId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"colorId": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					},
					"quantity": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"sizeId",
					"colorId",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Size": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateSizeModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProductCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"imgPath": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"name",
					"imgPath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateCategoryModel": {
				"properties": {
					"name": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"imgPath": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "token"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "belle-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LinhTran"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/report/products": {
			"get": {
				"operationId": "GetTopProductsReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProductReports"
								}
							}
						}
					}
				},
				"description": "Get list of products in descending  sales",
				"tags": [
					"Report Controller"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "time",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "(^[\\d]{4}$)|(^[\\d]{4}-([0][1-9]|[1][0-2])|(^week$))"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					}
				]
			}
		},
...
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProducts"
								}
							}
						}
					}
				},
				"description": "Get all products",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"in": "query",
						"name": "brand",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProductField"
						}
					},
					{
						"in": "query",
						"name": "change",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Change"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "min",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number",
							"minimum": 0
						}
					},
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProductInfo"
								}
							}
						}
					}
				},
				"description": "Create new product",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateProductModel"
							}
						}
					}
				}
			}
		},
		"/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProductInfo"
								}
							}
						}
					}
				},
				"description": "Get product details by its id",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProductInfo"
								}
							}
						}
					}
				},
				"description": "Update product info",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateProductModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProductById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete product",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
		"/products/variant": {
			"post": {
				"operationId": "CreateProductVariant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IVariantInfo"
								}
							}
						}
					}
				},
				"description": "Create product variant",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateVariantCreateModel"
							}
						}
					}
				}
			}
		},
		"/products/variant/{variantId}": {
			"patch": {
				"operationId": "UpdateProductVariant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IVariantInfo"
								}
							}
						}
					}
				},
				"description": "Update product variant",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "variantId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateVariantUpdateModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProductVariant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete product variant",
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "variantId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
		"/sizes": {
			"get": {
				"operationId": "GetSizes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Size"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all Sizes",
				"tags": [
					"Product Size"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateSize",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Size"
								}
							}
						}
					}
				},
				"description": "Create new Size",
				"tags": [
					"Product Size"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateSizeModel"
							}
						}
					}
				}
			}
		},
		"/sizes/{id}": {
			"patch": {
				"operationId": "UpdateSizeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Size"
								}
							}
						}
					}
				},
				"description": "Update Size info",
				"tags": [
					"Product Size"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateSizeModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSizeById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete Size",
				"tags": [
					"Product Size"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
		"/product-categories": {
			"get": {
				"operationId": "GetProductCategorys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all product categories",
				"tags": [
					"Product Category"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProductCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductCategory"
								}
							}
						}
					}
				},
				"description": "Create new product Category",
				"tags": [
					"Product Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateCategoryModel"
							}
						}
					}
				}
			}
		},
		"/product-categories/{id}": {
			"patch": {
				"operationId": "UpdateProductCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductCategory"
								}
							}
						}
					}
				},
				"description": "Update product category info",
				"tags": [
					"Product Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateCategoryModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProductCategoryById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete product category",
				"tags": [
					"Product Category"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}